+++ components/molecules/CardProfile/card-profile.tsx
@@ -7,6 +7,7 @@ import Text from "components/atoms/Typography/text";
 
 import ForkIcon from "img/icons/fork-icon.svg";
 import FirstPRIcon from "img/icons/first-pr-icon.svg";
+import Tooltip from "components/atoms/Tooltip/tooltip";
 
 interface CardProfileProps {
   githubAvatar?: string | StaticImageData;
@@ -28,13 +29,18 @@ const CardProfile = ({ githubAvatar, githubName, totalPRs, dateOfFirstPR }: Card
             <div className="flex text-xs items-center gap-1 text-light-slate-11">
               {totalPRs !== undefined && (
                 <>
-                  <Icon size={12} alt="Total Pull-Requests" IconImage={ForkIcon} /> {totalPRs} PR
+                  <Tooltip content="PRs merged">
+                    <Icon size={12} alt="PRs merged" IconImage={ForkIcon} />
+                  </Tooltip>
+                  {totalPRs} PR
                   {totalPRs === 1 ? "" : "s"}
                 </>
               )}
             </div>
             <div className="flex items-center gap-1 text-light-slate-11">
-              <Icon size={12} alt="Time from First Pull Request" IconImage={FirstPRIcon} /> {dateOfFirstPR}
+              <Tooltip content="First commit date">
+                <Icon size={12} alt="First commit date" IconImage={FirstPRIcon} /> {dateOfFirstPR}
+              </Tooltip>
             </div>
           </div>
         </div>