+++ components/organisms/TopNav/top-nav.tsx
@@ -17,23 +17,30 @@ const TopNav: React.FC = () => {
   const userInterest = gitHubUser?.interests.split(",")[0] || "javascript";
 
   return (
-    <header className="top-nav-container flex justify-between items-center px-2 md:px-16 py-0.5 bg-light-slate-3 border-b">
-      <div className="flex gap-8 items-center">
-        <HeaderLogo withBg={false} textIsBlack />
-        {!!user && onboarded ? (
-          <>
-            <Link className="text-sm text-light-slate-10" href={"/hub/insights"}>
-            Insights Hub
-            </Link>
-            <Link className="text-sm text-light-slate-10" href={`/${userInterest}/dashboard/filter/recent`}>
-            Explore
-            </Link>
-          </>
-        ) : (
-          ""
-        )}
+
+    <header className="top-nav-container flex justify-between items-center z-50 py-0.5 bg-light-slate-3 border-b">
+      <div className="flex justify-between items-center mx-auto container px-2 md:px-16">
+        <div className="flex gap-3 md:gap-8 items-center flex-wrap ">
+          <HeaderLogo withBg={false} textIsBlack />
+          {!!user && onboarded ? (
+            <>
+              <Link className="text-sm text-light-slate-10" href={"/hub/insights"}>
+                Insights Hub
+              </Link>
+              <Link className="text-sm text-light-slate-10" href={`/${userInterest}/dashboard/filter/recent`}>
+                Explore
+              </Link>
+            </>
+          ) : (
+            ""
+          )}
+          <Link className="text-sm text-light-slate-10" href={"/feed"}>
+            Highlights
+          </Link>
+        </div>
+        <AuthSection />
+
       </div>
-      <AuthSection />
     </header>
   );
 };